+++ s4e-backend/src/main/java/pl/cyfronet/s4e/data/repository/query/QueryText.java
@@ -17,18 +17,18 @@ public class QueryText extends QueryDecorator {
         paramsToDBField.put(POLARISATION, "polarisation");
         paramsToDBField.put(PROCESSING_LEVEL, "processing_level");
         paramsToDBField.put(SENSOR_MODE, "sensor_mode");
-        paramsToDBField.put("collection", "collection");
-        paramsToDBField.put("timeliness", "timeliness");
-        paramsToDBField.put("instrument", "instrument");
-        paramsToDBField.put("productLevel", "product_level");
+        paramsToDBField.put(COLLECTION, "collection");
+        paramsToDBField.put(TIMELINESS, "timeliness");
+        paramsToDBField.put(INSTRUMENT, "instrument");
+        paramsToDBField.put(PRODUCT_LEVEL, "product_level");
         paramsToDBField.put(RELATIVE_ORBIT_NUMBER, "relative_orbit_number");
-        paramsToDBField.put("absoluteOrbitNumber", "absolute_orbit_number");
+        paramsToDBField.put(ABSOLUTE_ORBIT_NUMBER, "absolute_orbit_number");
     }
 
     @Override
     public void doPrepareQueryAndParameters(Map<String, Object> params,
-                                                   List<Object> parameters,
-                                                   StringBuilder resultQuery) {
+                                            List<Object> parameters,
+                                            StringBuilder resultQuery) {
         paramsToDBField.entrySet().forEach(entry -> queryBy(entry, params, parameters, resultQuery));
     }
 
@@ -38,7 +38,19 @@ public class QueryText extends QueryDecorator {
                          StringBuilder resultQuery) {
         if (params.containsKey(entry.getKey())) {
             resultQuery.append(" AND metadata_content->>'" + entry.getValue() + "' = ? ");
-            parameters.add(String.valueOf(params.get(entry.getKey())));
+            if (params.containsKey(POLARISATION)) {
+                parameters.add(parsePolarisation(String.valueOf(params.get(entry.getKey()))));
+            } else {
+                parameters.add(String.valueOf(params.get(entry.getKey())));
+            }
         }
+    }
+
+    private String parsePolarisation(String s) {
+        String[] split = s.split("[+|\\s]+");
+        if (split.length > 1) {
+            return "Dual " + split[0] + "/" + split[1];
+        }
+        return split[0];
     }
 }